Refactor aws-lc-sys CMake build logic
