+++ aws-lc-sys/builder/main.rs
@@ -8,6 +8,8 @@ use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 
+use cmake_builder::CmakeBuilder;
+
 #[cfg(any(
     feature = "bindgen",
     not(any(
@@ -18,6 +20,7 @@ use std::process::Command;
     ))
 ))]
 mod bindgen;
+mod cmake_builder;
 
 pub(crate) fn get_aws_lc_include_path(manifest_dir: &Path) -> PathBuf {
     manifest_dir.join("aws-lc").join("include")
@@ -66,8 +69,6 @@ impl Default for OutputLibType {
                 // Only dynamic if the value is set and is a "negative" value
                 return OutputLibType::Dynamic;
             }
-
-            return OutputLibType::Static;
         }
         OutputLibType::Static
     }
@@ -83,7 +84,7 @@ impl OutputLibType {
 }
 
 impl OutputLib {
-    fn libname(self, prefix: Option<&str>) -> String {
+    fn libname(self, prefix: &Option<String>) -> String {
         let name = match self {
             OutputLib::Crypto => "crypto",
             OutputLib::Ssl => "ssl",
@@ -95,16 +96,6 @@ impl OutputLib {
             name.to_string()
         }
     }
-}
-
-fn artifact_output_dir(path: &Path) -> PathBuf {
-    path.join("build")
-        .join("artifacts")
-        .join(get_platform_output_path())
-}
-
-fn get_platform_output_path() -> PathBuf {
-    PathBuf::new()
 }
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -123,94 +114,6 @@ fn test_command(executable: &OsStr, args: &[&OsStr]) -> bool {
         return output.status.success();
     }
     false
-}
-
-fn find_cmake_command() -> Option<&'static OsStr> {
-    if test_command("cmake3".as_ref(), &["--version".as_ref()]) {
-        Some("cmake3".as_ref())
-    } else if test_command("cmake".as_ref(), &["--version".as_ref()]) {
-        Some("cmake".as_ref())
-    } else {
-        None
-    }
-}
-
-fn get_cmake_config(manifest_dir: &PathBuf) -> cmake::Config {
-    cmake::Config::new(manifest_dir)
-}
-
-fn prepare_cmake_build(manifest_dir: &PathBuf, build_prefix: String) -> cmake::Config {
-    let mut cmake_cfg = get_cmake_config(manifest_dir);
-
-    if ["powerpc64", "powerpc"]
-        .iter()
-        .any(|arch| target_arch().eq_ignore_ascii_case(arch))
-    {
-        cmake_cfg.define("ENABLE_EXPERIMENTAL_BIG_ENDIAN_SUPPORT", "1");
-    }
-
-    if OutputLibType::default() == OutputLibType::Dynamic {
-        cmake_cfg.define("BUILD_SHARED_LIBS", "1");
-    } else {
-        cmake_cfg.define("BUILD_SHARED_LIBS", "0");
-    }
-
-    let opt_level = env::var("OPT_LEVEL").unwrap_or_else(|_| "0".to_string());
-    if opt_level.ne("0") {
-        if opt_level.eq("1") || opt_level.eq("2") {
-            cmake_cfg.define("CMAKE_BUILD_TYPE", "relwithdebinfo");
-        } else {
-            cmake_cfg.define("CMAKE_BUILD_TYPE", "release");
-        }
-    } else {
-        cmake_cfg.define("CMAKE_BUILD_TYPE", "debug");
-    }
-
-    cmake_cfg.define("BORINGSSL_PREFIX", build_prefix);
-    let include_path = manifest_dir.join("generated-include");
-    cmake_cfg.define(
-        "BORINGSSL_PREFIX_HEADERS",
-        include_path.display().to_string(),
-    );
-
-    // Build flags that minimize our crate size.
-    cmake_cfg.define("BUILD_TESTING", "OFF");
-    if cfg!(feature = "ssl") {
-        cmake_cfg.define("BUILD_LIBSSL", "ON");
-    } else {
-        cmake_cfg.define("BUILD_LIBSSL", "OFF");
-    }
-    // Build flags that minimize our dependencies.
-    cmake_cfg.define("DISABLE_PERL", "ON");
-    cmake_cfg.define("DISABLE_GO", "ON");
-
-    if target_vendor() == "apple" {
-        if target_os().trim() == "ios" {
-            cmake_cfg.define("CMAKE_SYSTEM_NAME", "iOS");
-            if target().trim().ends_with("-ios-sim") {
-                cmake_cfg.define("CMAKE_OSX_SYSROOT", "iphonesimulator");
-            }
-        }
-        if target_arch().trim() == "aarch64" {
-            cmake_cfg.define("CMAKE_OSX_ARCHITECTURES", "arm64");
-        }
-    }
-
-    if cfg!(feature = "asan") {
-        env::set_var("CC", "/usr/bin/clang");
-        env::set_var("CXX", "/usr/bin/clang++");
-        env::set_var("ASM", "/usr/bin/clang");
-
-        cmake_cfg.define("ASAN", "1");
-    }
-
-    cmake_cfg
-}
-
-fn build_rust_wrapper(manifest_dir: &PathBuf) -> PathBuf {
-    prepare_cmake_build(manifest_dir, prefix_string() + "_")
-        .configure_arg("--no-warn-unused-cli")
-        .build()
 }
 
 #[cfg(any(
@@ -298,9 +201,12 @@ macro_rules! cfg_bindgen_platform {
     };
 }
 
-fn main() {
-    use crate::OutputLib::{Crypto, RustWrapper, Ssl};
+trait Builder {
+    fn check_dependencies(&self) -> Result<(), String>;
+    fn build(&self) -> Result<(), String>;
+}
 
+fn main() {
     let mut is_bindgen_required = cfg!(feature = "bindgen");
     let output_lib_type = OutputLibType::default();
 
@@ -320,13 +226,20 @@ fn main() {
         is_bindgen_required = true;
     }
 
-    check_dependencies();
-
     let manifest_dir = env::current_dir().unwrap();
     let manifest_dir = dunce::canonicalize(Path::new(&manifest_dir)).unwrap();
     let prefix = prefix_string();
+    let out_dir_str = env::var("OUT_DIR").unwrap();
+    let out_dir = Path::new(out_dir_str.as_str()).to_path_buf();
 
-    let out_dir = build_rust_wrapper(&manifest_dir);
+    let builder = CmakeBuilder::new(
+        manifest_dir.clone(),
+        out_dir.clone(),
+        Some(prefix.clone()),
+        output_lib_type,
+    );
+
+    builder.check_dependencies().unwrap();
 
     #[allow(unused_assignments)]
     let mut bindings_available = false;
@@ -360,31 +273,7 @@ fn main() {
         bindings_available,
         "aws-lc-sys build failed. Please enable the 'bindgen' feature on aws-lc-rs or aws-lc-sys"
     );
-
-    println!(
-        "cargo:rustc-link-search=native={}",
-        artifact_output_dir(&out_dir).display()
-    );
-
-    println!(
-        "cargo:rustc-link-lib={}={}",
-        output_lib_type.rust_lib_type(),
-        Crypto.libname(Some(&prefix))
-    );
-
-    if cfg!(feature = "ssl") {
-        println!(
-            "cargo:rustc-link-lib={}={}",
-            output_lib_type.rust_lib_type(),
-            Ssl.libname(Some(&prefix))
-        );
-    }
-
-    println!(
-        "cargo:rustc-link-lib={}={}",
-        output_lib_type.rust_lib_type(),
-        RustWrapper.libname(Some(&prefix))
-    );
+    builder.build().unwrap();
 
     println!(
         "cargo:include={}",
@@ -394,22 +283,6 @@ fn main() {
     println!("cargo:rerun-if-changed=builder/");
     println!("cargo:rerun-if-changed=aws-lc/");
     println!("cargo:rerun-if-env-changed=AWS_LC_SYS_STATIC");
-}
-
-fn check_dependencies() {
-    let mut missing_dependency = false;
-
-    if let Some(cmake_cmd) = find_cmake_command() {
-        env::set_var("CMAKE", cmake_cmd);
-    } else {
-        eprintln!("Missing dependency: cmake");
-        missing_dependency = true;
-    };
-
-    assert!(
-        !missing_dependency,
-        "Required build dependency is missing. Halting build."
-    );
 }
 
 fn setup_include_paths(out_dir: &Path, manifest_dir: &Path) -> PathBuf {
